fix cmd uitable wrap issue
